refactor jump table decoding
