+++ frontend/javascripts/libs/task_pool.ts
@@ -1,5 +1,6 @@
 import type { Saga, Task } from "oxalis/model/sagas/effect-generators";
 import { join, call, fork } from "typed-redux-saga";
+
 /*
   Given an array of async tasks, processTaskWithPool
   allows to execute at most ${poolSize} tasks concurrently.
@@ -11,8 +12,7 @@ export default function* processTaskWithPool(
 ): Saga<void> {
   const startedTasks: Array<Task<void>> = [];
   let isFinalResolveScheduled = false;
-  // @ts-expect-error ts-migrate(7034) FIXME: Variable 'error' implicitly has type 'any' in some... Remove this comment to see the full error message
-  let error = null;
+  let error: Error | null = null;
 
   // @ts-expect-error ts-migrate(7006) FIXME: Parameter 'fn' implicitly has an 'any' type.
   function* forkSafely(fn): Saga<void> {
@@ -22,7 +22,7 @@ export default function* processTaskWithPool(
     try {
       yield* call(fn);
     } catch (e) {
-      error = e;
+      error = e as Error;
     }
   }
 
@@ -34,7 +34,6 @@ export default function* processTaskWithPool(
         // awaited now together.
         // @ts-expect-error ts-migrate(2769) FIXME: No overload matches this call.
         yield* join(startedTasks);
-        // @ts-expect-error ts-migrate(7005) FIXME: Variable 'error' implicitly has an 'any' type.
         if (error != null) throw error;
       }
 
@@ -53,6 +52,7 @@ export default function* processTaskWithPool(
 
   for (let i = 0; i < poolSize; i++) {
     yield* fork(startNextTask);
-  } // The saga will wait for all forked tasks to terminate before returning, because
+  }
+  // The saga will wait for all forked tasks to terminate before returning, because
   // fork() creates attached forks (in contrast to spawn()).
 }
